+++ openeogeotrellis/geopysparkdatacube.py
@@ -20,9 +20,9 @@ import geopandas as gpd
 import pyproj
 import pytz
 import xarray as xr
-from geopyspark import TiledRasterLayer, Pyramid, Tile, SpaceTimeKey, SpatialKey, Metadata
+from geopyspark import TiledRasterLayer, Pyramid, Tile, SpaceTimeKey, SpatialKey, Metadata, zfactor_lat_lng_calculator
 from geopyspark.geotrellis import Extent, ResampleMethod
-from geopyspark.geotrellis.constants import CellType
+from geopyspark.geotrellis.constants import CellType, Unit
 from pandas import Series
 from pyproj import CRS
 from shapely.geometry import mapping, Point, Polygon, MultiPolygon, GeometryCollection, box
@@ -1945,7 +1945,7 @@ class GeopysparkDataCube(DriverDataCube):
 
             return latlng_extent.xmin, latlng_extent.ymin, latlng_extent.xmax, latlng_extent.ymax
 
-        def return_netcdf_items(java_items, bands, nodata) -> dict:
+        def return_netcdf_items(java_items, nodata) -> dict:
             items = {}
 
             for java_item in java_items:
@@ -1963,8 +1963,18 @@ class GeopysparkDataCube(DriverDataCube):
                         "bbox": bbox,
                         "geometry": geometry,
                     }
-                    if bands is not None:
-                        assets[asset_key]["bands"] = bands
+                    bands = []
+                    asset_metadata = asset.metadata()
+                    for band in asset_metadata.get("bands"):
+                        band = dict(band)
+                        if "statistics" in band:
+                            band["statistics"] = dict(band.get("statistics"))
+                        bands.append(band)
+                    assets[asset_key]["bands"] = bands
+                    assets[asset_key]["proj:bbox"] = tuple(asset_metadata.get("proj:bbox"))
+                    assets[asset_key]["proj:shape"] = tuple(asset_metadata.get("proj:shape"))
+                    assets[asset_key]["proj:epsg"] = asset_metadata.get("proj:epsg")
+                    assets[asset_key] = dict_no_none(assets[asset_key])
 
                 items[java_item.id()] = {
                     "id": java_item.id(),
@@ -2049,6 +2059,8 @@ class GeopysparkDataCube(DriverDataCube):
                 max_level = max_level.to_spatial_layer()
 
             if format == "GTIFF":
+                compression = format_options.get("compression", "deflate")
+                predictor = format_options.get("predictor", 1)
                 zlevel = format_options.get("ZLEVEL", 6)
                 tile_size = format_options.get("tile_size")
 
@@ -2075,6 +2087,7 @@ class GeopysparkDataCube(DriverDataCube):
 
                 if stitch:
                     gtiff_options = get_jvm().org.openeo.geotrellis.geotiff.GTiffOptions()
+                    gtiff_options.setCompression(compression, zlevel, predictor)
                     if filename_prefix.isDefined():
                         gtiff_options.setFilenamePrefix(filename_prefix.get())
                     gtiff_options.setResampleMethod(overview_resample)
@@ -2150,6 +2163,7 @@ class GeopysparkDataCube(DriverDataCube):
                 else:
                     _log.info("save_result: saveRDD")
                     gtiff_options = get_jvm().org.openeo.geotrellis.geotiff.GTiffOptions()
+                    gtiff_options.setCompression(compression, zlevel, predictor)
                     if filename_prefix.isDefined():
                         gtiff_options.setFilenamePrefix(filename_prefix.get())
                     gtiff_options.setSeparateAssetPerBand(separate_asset_per_band)
@@ -2381,6 +2395,7 @@ class GeopysparkDataCube(DriverDataCube):
             nodata = max_level.layer_metadata.no_data_value
             global_metadata = format_options.get("file_metadata",{})
             zlevel = format_options.get("ZLEVEL", 6)
+            add_bands_statistics = format_options.get("add_bands_statistics",False)
             for band_name, band_metadata in bands_metadata.items():
                 for tag, value in band_metadata.items():
                     bands_metadata[band_name][tag] = str(value)
@@ -2392,6 +2407,12 @@ class GeopysparkDataCube(DriverDataCube):
                     geometries = GeometryCollection(geometries.geoms)
                 projected_polygons = to_projected_polygons(get_jvm(), geometries)
                 labels = self.get_labels(geometries,feature_id_property)
+                options = get_jvm().org.openeo.geotrellis.netcdf.NetCDFOptions()
+                options.setBandNames(band_names)
+                options.setDimensionNames(dim_names)
+                options.setAttributes(global_metadata)
+                options.setBandsMetadata(bands_metadata)
+                options.setAddBandStatistics(add_bands_statistics)
                 if max_level.layer_type != gps.LayerType.SPATIAL:
                     _log.debug(f"projected_polygons carries {len(projected_polygons.polygons())} polygons")
                     java_items = get_jvm().org.openeo.geotrellis.netcdf.NetCDFRDDWriter.saveSamples(
@@ -2399,10 +2420,7 @@ class GeopysparkDataCube(DriverDataCube):
                         save_directory,
                         projected_polygons,
                         labels,
-                        band_names,
-                        dim_names,
-                        global_metadata,
-                        bands_metadata,
+                        options,
                         filename_prefix,
                     )
                 else:
@@ -2411,26 +2429,24 @@ class GeopysparkDataCube(DriverDataCube):
                         save_directory,
                         projected_polygons,
                         labels,
-                        band_names,
-                        dim_names,
-                        global_metadata,
-                        bands_metadata,
+                        options,
                         filename_prefix,
                     )
 
-                return return_netcdf_items(java_items, bands, nodata)
+                return return_netcdf_items(java_items, nodata)
             else:
                 originalName = pathlib.Path(filename)
                 filename_tmp = format_options.get("filename_prefix", "openEO") + ".nc" if originalName.name == "out" else originalName.name
                 if not stitch:
                     filename = save_directory + "/" + filename_tmp
+                    options = get_jvm().org.openeo.geotrellis.netcdf.NetCDFOptions()
+                    options.setBandNames(band_names)
+                    options.setDimensionNames(dim_names)
+                    options.setAttributes(global_metadata)
+                    options.setBandsMetadata(bands_metadata)
+                    options.setZLevel(zlevel)
+                    options.setAddBandStatistics(add_bands_statistics)
                     if strict_cropping:
-                        options = get_jvm().org.openeo.geotrellis.netcdf.NetCDFOptions()
-                        options.setBandNames(band_names)
-                        options.setDimensionNames(dim_names)
-                        options.setAttributes(global_metadata)
-                        options.setBandsMetadata(bands_metadata)
-                        options.setZLevel(zlevel)
                         options.setCropBounds(crop_extent)
                         java_items = get_jvm().org.openeo.geotrellis.netcdf.NetCDFRDDWriter.writeRasters(
                             max_level.srdd.rdd(),
@@ -2441,23 +2457,15 @@ class GeopysparkDataCube(DriverDataCube):
                             java_items = get_jvm().org.openeo.geotrellis.netcdf.NetCDFRDDWriter.saveSingleNetCDF(
                                 max_level.srdd.rdd(),
                                 filename,
-                                band_names,
-                                dim_names,
-                                global_metadata,
-                                bands_metadata,
-                                zlevel,
+                                options,
                             )
                         else:
                             java_items = get_jvm().org.openeo.geotrellis.netcdf.NetCDFRDDWriter.saveSingleNetCDFSpatial(
                                 max_level.srdd.rdd(),
                                 filename,
-                                band_names,
-                                dim_names,
-                                global_metadata,
-                                bands_metadata,
-                                zlevel,
+                                options,
                             )
-                    return return_netcdf_items(java_items, bands, nodata)
+                    return return_netcdf_items(java_items, nodata)
 
                 else:
                     if not tiled:
@@ -2942,6 +2950,28 @@ class GeopysparkDataCube(DriverDataCube):
             )
         )
         return atmo_corrected
+
+    @callsite
+    def aspect(self):
+        new_metadata = self.metadata
+        if self.metadata.has_band_dimension():
+            band_names = [band_name + '_aspect' for band_name in self.metadata.band_names]
+            new_metadata = self.metadata.with_new_band_names(band_names)
+        def compute_aspect(rdd, level):
+            pr = gps.get_spark_context()._jvm.org.openeo.geotrellis.OpenEOProcesses()
+            return pr.aspect(rdd)
+        return self._apply_to_levels_geotrellis_rdd(compute_aspect, metadata=new_metadata)
+
+    @callsite
+    def slope(self):
+        new_metadata = self.metadata
+        if self.metadata.has_band_dimension():
+            band_names = [band_name + '_slope' for band_name in self.metadata.band_names]
+            new_metadata = self.metadata.with_new_band_names(band_names)
+        def compute_slope(rdd, level):
+            pr = gps.get_spark_context()._jvm.org.openeo.geotrellis.OpenEOProcesses()
+            return pr.slope(rdd)
+        return self._apply_to_levels_geotrellis_rdd(compute_slope, metadata=new_metadata)
 
     def sar_backscatter(self, args: SarBackscatterArgs) -> 'GeopysparkDataCube':
         # Nothing to do: the actual SAR backscatter processing already happened in `load_collection`